feat(rss-bridge): Add Apache proxy to hide creds
The Spotify bridge in RSS-Bridge requires creds in the form of CLIENTID
and CLIENTSECRET in order to query the Spotify API. Because I want to
store in a public git repository, I want to avoid to hardcode such
credentials in all the feeds' URIs and then leak them.

I setup a web-server front-end which stores these creds in environment
variables (rss-bridge/rss-bridge-proxy/.env) and automatically add them
when needed.

IOW, this is just a workaround to avoid to store the creds in the feeds
URLs.